accessible/tests/mochitest/relations/test_embeds.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Embeds relation tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../relations.js"></script>
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
  <![CDATA[
    async function doTests() {
      testRelation(browserDocument(), RELATION_EMBEDS,
                   getAccessible(currentTabDocument()));

      enableLogging("docload");

      let loaded = waitForEvent(EVENT_DOCUMENT_LOAD_COMPLETE,
                                e => e.DOMNode == currentTabDocument());
      tabBrowser().loadURI(Services.io.newURI("about:robots"), {
        triggeringPrincipal:
          Services.scriptSecurityManager.getSystemPrincipal(),
      });
      await loaded;
      testRelation(browserDocument(), RELATION_EMBEDS,
                   getAccessible(currentTabDocument()));

      loaded = waitForEvent(EVENT_DOCUMENT_LOAD_COMPLETE,
                            e => e.DOMNode == currentTabDocument());
      tabBrowser().addTab("about:mozilla", {
        referrerURI: null,
        charset: null,
        postData: null,
        inBackground: false,
        triggeringPrincipal:
          Services.scriptSecurityManager.getSystemPrincipal(),
      });
      await loaded;
      testRelation(browserDocument(), RELATION_EMBEDS,
                   getAccessible(currentTabDocument()));

      disableLogging();
      closeBrowserWindow();
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    openBrowserWindow(doTests, "about:license");
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=707654"
       title="Embeds relation on root accessible can return not content document">
      Mozilla Bug 707654
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>
  </vbox>
</window>
